pyproject.toml
[project]
name = "financial-transactions-etl"
version = "0.1.0"
description = "ETL pipeline for financial transactions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.24.0",
    "kafka-python>=2.2.11",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "simplejson>=3.20.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
]
